Add osynic-osuapi.deno.dev Website (#71)
Hello everyone! (´｡• ω •｡`)

I'd like to humbly submit my little contribution to the awesome Leptos community. I've created a demo website for OsynicOsuapi using Leptos and would be honored if it could be added to the AwesomeLeptos collection.

The website demonstrates Leptos' capabilities combined with TailwindCSS, Lucide Leptos, Leptos Fluent, and Osynic Osuapi. It's a simple project, but I hope it can be useful or inspiring to others!

Thank you for considering my submission! ✨